clip/pyproject.toml
[project]
name = "clip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "pillow==10.3.0",
    "torch==2.3.1",
    "transformers==4.41.2",
]
